chore: cleanup structure
