+++ examples/asset-viewer-social-media/src/useCase/Authentication.ts
